+++ metrics/bow_mlp_test/metrics.txt
@@ -1,0 +1,17 @@
+2023-12-13 13:34:23.826510
+Accuracy: 0.67
+F1 positive label: 0.727
+F1 negative label: 0.582
+F1 mean score: 0.6552023-12-13 14:11:36.086759
+Accuracy: 0.67
+F1 positive label: 0.727
+F1 negative label: 0.582
+F1 mean score: 0.6552023-12-13 14:14:18.798725
+Accuracy: 0.804
+F1 positive label: 0.795
+F1 negative label: 0.812
+F1 mean score: 0.8042023-12-13 14:21:56.755986
+Accuracy: 0.804
+F1 positive label: 0.795
+F1 negative label: 0.812
+F1 mean score: 0.804
